Add balloons to SummaryTest.js
